Show match type reduction failures for other errors
Also show match type reduction failures for implicit search errors and member not found errors.
